Implementation of an algorithm to optimally choose employees to maximize total "value" in a employment tree.
In detail, this is the solution to the following prompt: Given a tree of
Employees, rooted at the CEO and with each employee having a boss and a
value, and a value n, choose n Employees to maximize total value, given
that:

Each employee adds its value plus the value of every boss in their chain
(boss, boss's boss, etc.)
Each employee can only provide its value to the total once
I used a greedy algorithm technique of choosing the leaf employee with
the greatest total value at each step and adjusting the total values of
the employees affected by this choice. The leaf nodes were stored in a
Priority Queue for efficient access to the maximum value node (O(1)) and
readjustment (O(log n)).